Initial ver with networktables4 and megatag2 integration

+++ src/main/java/ca/team1310/swerve/SwerveTelemetry.java
@@ -134,11 +134,11 @@ public final class SwerveTelemetry {
     /**
      * The priority id from the vision system
      */
-    public double visionPriorityId = Double.MIN_VALUE;
+    public long visionPriorityId = Long.MIN_VALUE;
     /**
      * The tid value from the vision system
      */
-    public double visionTid = Double.MIN_VALUE;
+    public long visionTid = Long.MIN_VALUE;
     /**
      * The tx value from the vision system
      */
